+++ .github/update_changelog.py
@@ -122,9 +122,8 @@ new_entry = re.sub(
 new_entry = re.sub(
     r"\n\n.*",
     "\n\n"
-    "```{important}\n"
-    "This version is not released yet and is under active development.\n"
-    "```\n\n",
+    "> [!IMPORTANT]\n"
+    "> This version is not released yet and is under active development.\n\n",
     new_entry,
     flags=re.MULTILINE | re.DOTALL,
 )